+++ src/js/core/height-viewport.js
@@ -15,7 +15,7 @@ export default function (UIkit) {
         },
 
         init() {
-            this._callUpdate();
+            this.$emit();
         },
 
         update: {