+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add support for 'data-uk-' prefixed component attributes
 - Add support for Primary Arguments in components
 - Add support for Functional Components
+- Components reinitialize on added/removed children
 
 ### Fixed